fix(rocJPEG): widen GetChannelPitchAndSizes channel sizes to size_t
GetChannelPitchAndSizes() computed each channel size as uint32_t
pitch * height and stored it into a uint32_t array. pitch and height
derive from the JPEG SOF marker (up to 65535 each), so for large images
the product exceeds 32 bits and wraps. Callers hipMalloc() the device
channel buffers from those wrapped (undersized) values while SaveImage()
now copies the full size_t surface size, giving an out-of-bounds read on
the device buffer -- the mismatch MythreyaK flagged in review.

Perform the size arithmetic in size_t and store channel_sizes as size_t
so the device allocation and the host-side copy agree. The two
rocprofiler copies wrapped the product in align(int, int), which
overflows before aligning, so add a size_t align_size() helper for the
channel-size math there.

Updated in all three vendored copies and their callers:
- projects/rocjpeg/samples (jpegDecode, jpegDecodeBatched, jpegDecodePerf)
- projects/rocprofiler-sdk/tests/bin/rocjpeg
- projects/rocprofiler-systems/examples/jpegdecode

JIRA ID: SWSPLAT-24294

+++ projects/rocjpeg/samples/jpegDecode/jpegdecode.cpp
@@ -28,8 +28,8 @@ int main(int argc, char **argv) {
     uint8_t num_components;
     uint32_t widths[ROCJPEG_MAX_COMPONENT] = {};
     uint32_t heights[ROCJPEG_MAX_COMPONENT] = {};
-    uint32_t channel_sizes[ROCJPEG_MAX_COMPONENT] = {};
-    uint32_t prior_channel_sizes[ROCJPEG_MAX_COMPONENT] = {};
+    size_t channel_sizes[ROCJPEG_MAX_COMPONENT] = {};
+    size_t prior_channel_sizes[ROCJPEG_MAX_COMPONENT] = {};
     uint32_t num_channels = 0;
     int total_images = 0;
     double time_per_image_all = 0;
